Reliably register button presses
The design of stm32f0xx_hal seems antithetical to sensible use of
interrupts so this was much harder than it should have been. In short,
it's not possible to clear peripherals status bits in interrupt handlers
since the library's various abstractions (e.g. `Timer`) expect to do so
themselves. This essentially precludes use of interrupts since otherwise
you will loop in the handler.